# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Intelligence Platform",
    "description": "FastAPI Backend for N100 Financial Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Company Annual Reports & Documents",
        "description": "Retrieve annual report links and link validity status for requested company.",
        "operationId": "get_company_documents_api_v1_companies__ticker__documents_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document list returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDocumentsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Company ticker not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompanyDocumentsResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker",
            "description": "Requested company ticker symbol"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Official company name"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "description": "Total annual report documents found"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "type": "array",
            "title": "Documents",
            "description": "List of annual report document entries"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "company_name",
          "document_count",
          "documents"
        ],
        "title": "CompanyDocumentsResponse"
      },
...
      "DocumentItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Primary key ID"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id",
            "description": "Company ticker symbol"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "description": "Financial report year"
          },
          "annual_report": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Report",
            "description": "Annual report document URL"
          },
          "document_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Url",
            "description": "Alias for annual report URL"
          },
          "is_url_valid": {
            "type": "boolean",
            "title": "Is Url Valid",
            "description": "Boolean flag indicating whether document URL is valid"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "is_url_valid"
        ],
        "title": "DocumentItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}